[level 0] Title: 9로 나눈 나머지, Time: 6.36 ms, Memory: 93.9 MB -BaekjoonHub

+++ 프로그래머스/0/181914. 9로 나눈 나머지/9로 나눈 나머지.java
@@ -3,7 +3,7 @@ class Solution {
         int sum = 0;
         
         for(char num: number.toCharArray()) {
-            sum += (int)(num - '0');
+            sum += num - '0';
         }
         
         return sum % 9;
